+++ jest.config.js
@@ -28,4 +28,6 @@ module.exports = {
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
   },
+
+  errorOnDeprecated: true,
 };