+++ packages/cli/src/typings.d.ts
@@ -1,7 +1,5 @@
 declare module 'is-valid-path' {
-  namespace isValidPath {
-
-  }
+  namespace isValidPath {}
   function isValidPath(val: string): boolean;
   export = isValidPath;
 }