+++ packages/core/__tests__/diff/enum.ts
@@ -1,145 +1,145 @@
-import {buildASTSchema} from 'graphql';
-import gql from 'graphql-tag';
+import {buildSchema} from 'graphql';
 
+import {findFirstChangeByPath} from '../../utils/testing';
 import {diff} from '../../src/index';
 import {CriticalityLevel} from '../../src/diff/changes/change';
 
-test('enum with new value', () => {
-  const schemaA = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+describe('enum', () => {
+  test('value added', () => {
+    const a = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A
-      B
-    }
-  `);
+      enum enumA {
+        A
+        B
+      }
+    `);
 
-  const schemaB = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+    const b = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A
-      B
-      C
-    }
-  `);
+      enum enumA {
+        A
+        B
+        C
+      }
+    `);
 
-  const changes = diff(schemaA, schemaB);
+    const changes = diff(a, b);
+    const change = findFirstChangeByPath(changes, 'enumA.C');
 
-  expect(changes.length).toEqual(1);
-  expect(changes[0].criticality.level).toEqual(CriticalityLevel.Dangerous);
-  expect(changes[0].criticality.reason).toBeDefined();
-  expect(changes[0].message).toEqual(
-    `Enum value 'C' was added to enum 'enumA'`,
-  );
-  expect(changes[0].path).toEqual('enumA.C');
-});
+    expect(changes.length).toEqual(1);
+    expect(change.criticality.level).toEqual(CriticalityLevel.Dangerous);
+    expect(change.criticality.reason).toBeDefined();
+    expect(change.message).toEqual(`Enum value 'C' was added to enum 'enumA'`);
+  });
 
-test('enum with removed value', () => {
-  const schemaA = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+  test('value removed', () => {
+    const a = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A
-      B
-    }
-  `);
+      enum enumA {
+        A
+        B
+      }
+    `);
 
-  const schemaB = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+    const b = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A
-    }
-  `);
+      enum enumA {
+        A
+      }
+    `);
 
-  const changes = diff(schemaA, schemaB);
+    const changes = diff(a, b);
+    const change = findFirstChangeByPath(changes, 'enumA.B');
 
-  expect(changes.length).toEqual(1);
-  expect(changes[0].criticality.level).toEqual(CriticalityLevel.Breaking);
-  expect(changes[0].criticality.reason).toBeDefined();
-  expect(changes[0].message).toEqual(
-    `Enum value 'B' was removed from enum 'enumA'`,
-  );
-  expect(changes[0].path).toEqual('enumA.B');
-});
+    expect(changes.length).toEqual(1);
+    expect(change.criticality.level).toEqual(CriticalityLevel.Breaking);
+    expect(change.criticality.reason).toBeDefined();
+    expect(change.message).toEqual(
+      `Enum value 'B' was removed from enum 'enumA'`,
+    );
+  });
 
-test('enum with description changed', () => {
-  const schemaA = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+  test('description changed', () => {
+    const a = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    """
-    Old Description
-    """
-    enum enumA {
-      A
-      B
-    }
-  `);
+      """
+      Old Description
+      """
+      enum enumA {
+        A
+        B
+      }
+    `);
 
-  const schemaB = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+    const b = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    """
-    New Description
-    """
-    enum enumA {
-      A
-      B
-    }
-  `);
+      """
+      New Description
+      """
+      enum enumA {
+        A
+        B
+      }
+    `);
 
-  const changes = diff(schemaA, schemaB);
+    const changes = diff(a, b);
+    const change = findFirstChangeByPath(changes, 'enumA');
 
-  expect(changes.length).toEqual(1);
-  expect(changes[0].criticality.level).toEqual(CriticalityLevel.NonBreaking);
-  expect(changes[0].message).toEqual(
-    `Description 'Old Description' on type 'enumA' has changed to 'New Description'`,
-  );
-  expect(changes[0].path).toEqual('enumA');
-});
+    expect(changes.length).toEqual(1);
+    expect(change.criticality.level).toEqual(CriticalityLevel.NonBreaking);
+    expect(change.message).toEqual(
+      `Description 'Old Description' on type 'enumA' has changed to 'New Description'`,
+    );
+  });
 
-test('enum with deprecationReason changed', () => {
-  const schemaA = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+  test('deprecation reason changed', () => {
+    const a = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A @deprecated(reason: "Old Reason")
-      B
-    }
-  `);
+      enum enumA {
+        A @deprecated(reason: "Old Reason")
+        B
+      }
+    `);
 
-  const schemaB = buildASTSchema(gql`
-    type Query {
-      fieldA: String
-    }
+    const b = buildSchema(/* GraphQL */ `
+      type Query {
+        fieldA: String
+      }
 
-    enum enumA {
-      A @deprecated(reason: "New Reason")
-      B
-    }
-  `);
+      enum enumA {
+        A @deprecated(reason: "New Reason")
+        B
+      }
+    `);
 
-  const changes = diff(schemaA, schemaB);
+    const changes = diff(a, b);
+    const change = findFirstChangeByPath(changes, 'enumA.A');
 
-  expect(changes.length).toEqual(1);
-  expect(changes[0].criticality.level).toEqual(CriticalityLevel.NonBreaking);
-  expect(changes[0].message).toEqual(
-    `Enum value 'enumA.A' deprecation reason changed from 'Old Reason' to 'New Reason'`,
-  );
-  expect(changes[0].path).toEqual('enumA.A');
+    expect(changes.length).toEqual(1);
+    expect(change.criticality.level).toEqual(CriticalityLevel.NonBreaking);
+    expect(change.message).toEqual(
+      `Enum value 'enumA.A' deprecation reason changed from 'Old Reason' to 'New Reason'`,
+    );
+  });
 });