+++ packages/core/src/diff/common/compare.ts
@@ -1,7 +1,11 @@
-export function equal<T>(a: T, b: T): boolean {
+export function isEqual<T>(a: T, b: T): boolean {
   return a === b || (!a && !b);
 }
 
-export function notEqual<T>(a: T, b: T): boolean {
-  return !equal(a, b);
+export function isNotEqual<T>(a: T, b: T): boolean {
+  return !isEqual(a, b);
+}
+
+export function isVoid<T>(a: T): boolean {
+  return typeof a === 'undefined' || a === null;
 }