+++ packages/core/src/diff/enum.ts
@@ -1,6 +1,6 @@
 import {GraphQLEnumType, GraphQLEnumValue} from 'graphql';
 
-import {notEqual} from './common/compare';
+import {isNotEqual} from './common/compare';
 import {
   enumValueRemoved,
   enumValueAdded,
@@ -33,11 +33,11 @@ export function changesInEnum(
   changes.push(...removed.map(v => enumValueRemoved(oldEnum, v)));
 
   common.forEach(({oldValue, newValue}) => {
-    if (notEqual(oldValue.description, newValue.description)) {
+    if (isNotEqual(oldValue.description, newValue.description)) {
       changes.push(enumValueDescriptionChanged(newEnum, oldValue, newValue));
     }
 
-    if (notEqual(oldValue.deprecationReason, newValue.deprecationReason)) {
+    if (isNotEqual(oldValue.deprecationReason, newValue.deprecationReason)) {
       changes.push(
         enumValueDeprecationReasonChanged(newEnum, oldValue, newValue),
       );